+++ Estimation/BamProcessing/BamController.h
@@ -36,7 +36,7 @@ namespace Estimation
 			const bool _gene_in_chromosome_name;
 			const std::string _read_param_filenames;
 			const std::string _gtf_path;
-			const int _min_barcode_quality;
+			const char _min_barcode_phred;
 
 		private:
 			void parse_bam_file(const std::string &bam_name, std::shared_ptr<BamProcessorAbstract> &processor,
@@ -60,7 +60,7 @@ namespace Estimation
 			                              const CellsDataContainer &container) const;
 
 			BamController(const BamTags &tags, bool filled_bam, const std::string &read_param_filenames,
-			              const std::string &gtf_path, bool gene_in_chromosome_name, int min_barcode_quality);
+			              const std::string &gtf_path, bool gene_in_chromosome_name, char min_barcode_phred);
 		};
 	}
 }