+++ Estimation/BamProcessing/FilledBamParamsParser.h
@@ -10,11 +10,11 @@ namespace BamProcessing
 	class FilledBamParamsParser : public ReadParamsParser
 	{
 	protected:
-		const int _min_barcode_quality;
+		const char _min_barcode_phred;
 
 	public:
 		FilledBamParamsParser(const std::string &gtf_path, const BamTags &tags, bool gene_in_chromosome_name,
-		                      int min_barcode_quality);
+		                      char min_barcode_phred);
 		bool get_read_params(const BamTools::BamAlignment &alignment, Tools::ReadParameters &read_params) override;
 	};
 }