+++ Tests/TestTools.cpp
@@ -71,7 +71,7 @@ BOOST_AUTO_TEST_SUITE(TestTools)
 		BOOST_CHECK_EQUAL(rp.cell_barcode(), "ATTTG");
 		BOOST_CHECK_EQUAL(rp.umi(), "ATAT");
 
-		rp = ReadParameters("AAATTTTATA", "TTGG", "QUALCB", "CCC", 0);
+		rp = ReadParameters("AAATTTTATA", "TTGG", "QUALCB", "CCC");
 		auto rp_info = ReadParameters::parse_from_string(rp.to_string("ID"));
 		BOOST_CHECK_EQUAL(rp_info.first, "ID");
 		BOOST_CHECK_EQUAL(rp_info.second.cell_barcode(), rp.cell_barcode());